server/requirements.txt
grpcio
grpcio-tools
psycopg2-binary
pytest
pytest-mock
